move AgentToolCallMessage into its own file and add copy buttons + tool call id to modal
